backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.72.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
dataclasses-json==0.6.7
distro==1.9.0
docstring_parser==0.17.0
ecdsa==0.19.1
fastapi==0.109.0
ffmpeg-python==0.2.0
openai-whisper==20231117
frozenlist==1.8.0
future==1.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.1.6
langchain-anthropic==0.1.1
langchain-community==0.0.20
langchain-core==0.1.23
langchain-openai==0.0.5
langgraph==0.0.20
langsmith==0.0.87
marshmallow==3.26.1
multidict==6.7.0
mypy_extensions==1.1.0
numpy==1.26.4
openai==1.109.1
packaging==23.2
pillow==10.2.0
propcache==0.4.1
pyasn1==0.6.1
pycparser==2.23
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.35.1
tenacity==8.5.0
tiktoken==0.5.2
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.27.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
